fix(app/worker/profile): refactor according to review
